move selected resource page to be the only resource page
